+++ CHANGELOG.md
@@ -28,6 +28,7 @@ The types of changes are:
   * Skeleton of page to add a system manually [#1068](https://github.com/ethyca/fides/pull/1068)
   * Refactor config wizard system forms to be reused for system management [#1072](https://github.com/ethyca/fides/pull/1072)
   * Add additional optional fields to system management forms [#1082](https://github.com/ethyca/fides/pull/1082)
+  * Delete a system through the UI [#1085](https://github.com/ethyca/fides/pull/1085)
 
 ### Changed